Filed GenTagNum and GenMassyMeat features into separate traits+files
