+++ grvx_mvn/grvx_zpol/src/main/scala/fun/gravax/zdynamo/GenMeatStuff.scala
@@ -1,0 +1,103 @@
+package fun.gravax.zdynamo
+
+import zio.{Chunk, NonEmptyChunk, RIO, UIO, ZIO, Random => ZRandom}
+import zio.stream.{UStream, ZStream}
+
+import java.math.MathContext
+import scala.collection.immutable.{Queue, Map => SMap}
+
+private trait GenMeatStuff
+
+trait GenMeatAndMassData extends KnowsBinItem {
+
+	def gaussianBD(zrnd : ZRandom, mathCtx : MathContext)(mean : BigDecimal, stdDev : BigDecimal): UIO[BigDecimal] = {
+		val stdNumEff: UIO[Double] = zrnd.nextGaussian
+		stdNumEff.map(num => {
+			// https://blogs.oracle.com/javamagazine/post/four-common-pitfalls-of-the-bigdecimal-class-and-how-to-avoid-them
+			val stdBD = BigDecimal(num, mathCtx)
+			val scaledAndShiftedBD = stdBD.*(stdDev).+(mean)
+			scaledAndShiftedBD
+		})
+	}
+
+	def truncGaussBD(zrnd : ZRandom, mathCtx : MathContext)(min : BigDecimal, max : BigDecimal): UIO[BigDecimal] = {
+		val diff = max.-(min)
+		val mean = min.+(diff./(2))
+		val stdDev = diff./(3)
+		val gbdEff = gaussianBD(zrnd, mathCtx)(mean, stdDev)
+		val bgbdEff : UIO[BigDecimal]  = gbdEff.repeatWhile(rslt => (rslt.compareTo(min) < 0 || rslt.compareTo(max) > 0))
+		bgbdEff
+	}
+
+	val (meanMin, meanMax) = (BigDecimal("-1.0"), BigDecimal("1.0"))
+	val (vrMin, vrMax) = (BigDecimal("0.0"), BigDecimal("0.5"))
+	def genRandStatEntry(zrnd : ZRandom, mathCtx : MathContext)(ekey : BinTypes.EntryKey) : UIO[BinTypes.StatEntry] = {
+		for {
+			mean <- truncGaussBD(zrnd, mathCtx)(meanMin, meanMax)
+			vr <- truncGaussBD(zrnd, mathCtx)(vrMin, vrMax)
+		} yield (ekey, mean, vr)
+	}
+
+
+	def genRandStatMap(zrnd : ZRandom, mathCtx : MathContext)(keys : Iterable[BinTypes.EntryKey]) : UIO[BinTypes.StatMap] = {
+		val kstrm: UStream[BinTypes.EntryKey] = ZStream.fromIterable(keys)
+		// This could be done in fewer lines using ZSink.collectAllToMap
+		val stStrm: UStream[(BinTypes.EntryKey, BinTypes.StatEntry)] = kstrm.flatMap(key => {
+			val steOp  = genRandStatEntry(zrnd, mathCtx)(key)
+			val tupOp = steOp.map(statEnt => (statEnt._1, statEnt))
+			ZStream.fromZIO(tupOp)
+		})
+		// val chunkMkrSink = ZSink.collectAll[(BinTypes.EntryKey, BinTypes.StatEntry)]
+		val stChunk_op: UIO[Chunk[(BinTypes.EntryKey, BinTypes.StatEntry)]] = stStrm.runCollect
+		val stMap: UIO[SMap[BinTypes.EntryKey, BinTypes.StatEntry]] = stChunk_op.map(chnk => chnk.toMap)
+		stMap
+	}
+	def genRandEntryKey(zrnd : ZRandom, keyLen : Int) : UIO[BinTypes.EntryKey] = {
+		val (minKeyChr, maxKeyChr) = ('A', 'Z')
+		val oneChrOp: UIO[Char] = zrnd.nextIntBetween(minKeyChr, maxKeyChr + 1).map(_.toChar)
+		val manyChrStrm = ZStream.repeatZIO(oneChrOp)
+		val enoughChrStrm = manyChrStrm.take(keyLen)
+		// val acs = ZSink.collectAll[Char]
+		val echk: UIO[Chunk[Char]] = enoughChrStrm.runCollect
+		val esop: UIO[String] = echk.map(chnk => new String(chnk.toArray))
+		esop
+	}
+	def genManyEKeys (zrnd : ZRandom, keyLen : Int, numKeys : Int) : UIO[Seq[BinTypes.EntryKey]] = {
+		val randEkeyOp = genRandEntryKey(zrnd, keyLen)
+		val keyStrm = ZStream.repeatZIO(randEkeyOp)
+		val enoughKeysOp = keyStrm.take(numKeys).runCollect.map(_.toSeq)
+		enoughKeysOp
+	}
+	// Generate a stream of random stat maps, and treat each as the meat-map of a bin.
+	def genMeatInfoStrmFromFixedKeys(zrnd : ZRandom, mathCtx : MathContext, fixedKeys : Seq[BinTypes.EntryKey], fixedFlavor : String): UStream[BinMeatInfo] = {
+		val oneStMpOp = genRandStatMap(zrnd, mathCtx)(fixedKeys)
+		val stMpStrm = ZStream.repeatZIO(oneStMpOp)
+		val miStrm = stMpStrm.map(stMap => BinMeatInfo(fixedFlavor, stMap))
+		miStrm
+	}
+
+	val (massMin, massMax) = (BigDecimal("1.0"), BigDecimal("1000.0"))
+	val (myKeyLen, myNumKeys) = (3, 50)
+
+	def mkMassGenOp(zrnd: ZRandom, mathCtx: MathContext): UIO[BigDecimal] = truncGaussBD(zrnd, mathCtx)(massMin, massMax)
+	def addMassToMeatStrm(meatStrm : UStream[BinMeatInfo], massGenOp : UIO[BigDecimal]) : UStream[(BigDecimal, BinMeatInfo)] = {
+		meatStrm.mapZIO(bmi => massGenOp.map(massBD => (massBD, bmi)))
+	}
+
+	// The configuration of this stream is sprinkled in the vals, currently Stu counts 8 of em
+	// Each stream tuple is ready to be the genesis of a leaf-bin (with no kiddos)
+	def mkMassyMeatStrm(zrnd: ZRandom, mathCtx: MathContext) : UStream[(BigDecimal, BinMeatInfo)] = {
+
+		val binFlavor = BFLV_ANN_RET_MEAN_VAR
+		val ekeysOp : UIO[Seq[BinTypes.EntryKey]] = genManyEKeys(zrnd, myKeyLen, myNumKeys)
+		val ekeyStrmOfSeq = ZStream.fromZIO(ekeysOp)
+		val massyMeatStrm : UStream[(BigDecimal, BinMeatInfo)] = ekeyStrmOfSeq.flatMap(keySeq => {
+			val meatInfoStrm = genMeatInfoStrmFromFixedKeys(zrnd, mathCtx, keySeq, binFlavor)
+			val massOp = mkMassGenOp(zrnd, mathCtx)
+			val massyMeatStrm = addMassToMeatStrm(meatInfoStrm, massOp)
+			massyMeatStrm
+		})
+		massyMeatStrm
+	}
+
+}
